Merged upstream & first TreckingBike Profile

+++ mgmap/src/main/java/mg/mgmap/activity/mgmap/features/routing/CostCalculator.java
@@ -2,7 +2,7 @@ package mg.mgmap.activity.mgmap.features.routing;
 
 public interface CostCalculator {
 
-    double calcCosts(double dist, double verDist);
+    double calcCosts(double dist, float verDist);
 
     double heuristic( double dist, float vertDist);
 }
